+++ bc-cache.php
@@ -4,10 +4,10 @@
  * Plugin Name: BC Cache
  * Plugin URI: https://github.com/chesio/bc-cache
  * Description: Simple full page cache plugin inspired by Cachify.
- * Version: 2.0.2
+ * Version: 2.1.0
  * Author: Česlav Przywara <ceslav@przywara.cz>
  * Author URI: https://www.chesio.com
- * Requires PHP: 7.2
+ * Requires PHP: 7.3
  * Requires WP: 5.5
  * Tested up to: 5.8
  * Text Domain: bc-cache
@@ -15,28 +15,65 @@
  * Update URI: https://github.com/chesio/bc-cache
  */
 
-if (version_compare(PHP_VERSION, '7.2', '<')) {
-    // Warn user that his/her PHP version is too low for this plugin to function.
-    add_action('admin_notices', function () {
-        echo '<div class="error"><p>';
-        echo esc_html(
-            sprintf(
-                __('BC Cache plugin requires PHP 7.2 to function properly, but you have version %s installed. The plugin has been auto-deactivated.', 'bc-cache'),
-                PHP_VERSION
-            )
-        );
-        echo '</p></div>';
-        // https://make.wordpress.org/plugins/2015/06/05/policy-on-php-versions/
-        if (isset($_GET['activate'])) {
-            unset($_GET['activate']);
+// Check plugin requirements, disable if they are not met.
+if (
+    false === call_user_func(
+        function () {
+            $php_version_ok = version_compare(PHP_VERSION, '7.3', '>=');
+            $pretty_permalinks_on = (bool) get_option('permalink_structure');
+
+            if (!$php_version_ok) {
+                // Warn user that his/her PHP version is too low for this plugin to function.
+                add_action('admin_notices', function () {
+                    echo '<div class="error"><p>';
+                    echo esc_html(
+                        sprintf(
+                            __('BC Cache plugin requires PHP 7.3 to function properly, but you have version %s installed. The plugin has been auto-deactivated.', 'bc-cache'),
+                            PHP_VERSION
+                        )
+                    );
+                    echo '</p></div>';
+                }, 10, 0);
+            }
+
+            if (!$pretty_permalinks_on) {
+                // Warn user that plugin does not work without pretty permalinks activated.
+                add_action('admin_notices', function () {
+                    echo '<div class="error"><p>';
+                    echo sprintf(
+                        __('BC Cache plugin requires %s to be activated, but your website have them turned off. The plugin has been auto-deactivated.', 'bc-cache'),
+                        sprintf(
+                            '<a href="%s">%s</a>',
+                            admin_url('options-permalink.php'),
+                            esc_html('pretty permalinks', 'bc-cache')
+                        )
+                    );
+                    echo '</p></div>';
+                }, 10, 0);
+            }
+
+            if (!$php_version_ok || !$pretty_permalinks_on) {
+                add_action('admin_notices', function () {
+                    // https://make.wordpress.org/plugins/2015/06/05/policy-on-php-versions/
+                    if (isset($_GET['activate'])) {
+                        unset($_GET['activate']);
+                    }
+                }, 10, 0);
+
+                // Self deactivate.
+                add_action('admin_init', function () {
+                    deactivate_plugins(plugin_basename(__FILE__));
+                }, 10, 0);
+
+                // Requirements check failed.
+                return false;
+            }
+
+            // Requirements check passed.
+            return true;
         }
-    }, 10, 0);
-
-    // Self deactivate.
-    add_action('admin_init', function () {
-        deactivate_plugins(plugin_basename(__FILE__));
-    }, 10, 0);
-
+    )
+) {
     // Bail.
     return;
 }