+++ tests/static-analysis/bootstrap.php
@@ -1,7 +1,4 @@
 <?php
 
-define('WP_CONTENT_DIR', '');
-define('WP_CONTENT_URL', '');
-define('WP_CRON_LOCK_TIMEOUT', 60);
-define('COOKIEPATH', '');
-define('COOKIE_DOMAIN', '');
+require_once __DIR__ . '/stubs/autodescription.php';
+require_once __DIR__ . '/stubs/wordpress.php';