+++ tests/static-analysis/stubs/wordpress.php
@@ -1,0 +1,7 @@
+<?php
+
+define('WP_CONTENT_DIR', '');
+define('WP_CONTENT_URL', '');
+define('WP_CRON_LOCK_TIMEOUT', 60);
+define('COOKIEPATH', '');
+define('COOKIE_DOMAIN', '');